move unix-only network install stuff under netinst
the Dagger CLI and shim depends on network now for some shared consts,
so we need to keep its dependencies tidy and cross-platform
