+++ network/netinst/paths.go
@@ -1,4 +1,4 @@
-package network
+package netinst
 
 import "fmt"
