+++ network/netinst/util.go
@@ -1,4 +1,4 @@
-package network
+package netinst
 
 import (
 	"os"
